Kill all the routing and emit instead
the only reason I wanted to keep update being sent up and down the tree was so that parent components could communicate with it, doing an elm-like send-extra-stuff get-extra-stuff call to render.

I'm pretty sure the only interesting things in send-extra/get-extra calls to child updates are the extra parts -- parents should _never_ screw around with what's being unwrapped and sent, and should never screw around with the
main return value that should become that child's next state.

So, I decided to just take that responsibility out of parent components' hands, and just give them a function to "yield" to the child, optionally sending in data, and optionally doing stuff with the return.

This ended up being almost cool, but sucks because it introduces new concepts. update() becomes more complicated with weird function signatures, and... stuff.

SO instead, I scrapped all of that and made things really really simple:

1. Children are barely managed _at all_ by parents. WOOOOOOO!!!! No need to route _anything_ to and from them. AUTOMATIC PLUMBING YEAH
2. Passing data down to children by props a la react is restored, _but with a twist_: props are available to the child in `init` and in `update` (making extra stuff in up date elm-ish), but not in the view. I think this makes
sense!!! I hope it makes sense!!!
3. Children can pass data back up similar to elm by adding extra stuff to update()'s return, buuut _(twist! again!)_, the parent opts in to getting that stuff by passing `onEmit` in the view, giving it a Msg tag, like any other
onEvent handler. So, when children return extra stuff, we then call the parent updater with the Msg tag it assigned to the child, which it can use to handle that new data. This is hard to type out concisely!

This is starting to feel pretty right. Holy moly it's lost _sooooooo_ much of elm by losing all the plumbing, but HOLY MOLY NO PLUMBING

- There is still a _little_ bit of parent bookkeeping required to handle the n-components-of-type case. I won't be dead if this doesn't go away. It feels like it should be able to though, with some clever use of the `key`
prop perhaps.

- I've killed the route component for now. A route component is still required: ultimately we do have to aggregate effects and subscriptions. I'm hoping that, using React's `context`, the top-most component can detect that it's
the root and set that up for itself, and provide stuff for children to do effects and subscriptions (with no plumbing!) via a context it provides. We'll seeeeeee.....