Get context to work with init
Context is not available in `constructor`, but we can't call `init` until we have `spindle` from `context`. We can't render until `init`, so, this commit hacks the lifecycle to render `null` until we've set state from
`componentDidMount`, where we can do all the context stuff.

This is annoying, and creates a case where if a child doesn't implement `init` correctly in a component near the root, children of it may never be mounted and just sliently disappear.

Hopefully cases like that can be detected to give good dev warnings.

Hopefully the cascade of renders down the tree as each component can render and _then_ mount its children who have to init and can then render... doesn't matter too much.

:(