Add a command example: dice roll
Writing this example ran me straight into react's surprising `setState` behaviour: in some contexts, `setState` will push a queued state update to apply after the event handler (or lifecycle method or whatever) has finished
running, and reading `this.state` will show the old value. In other cases, it will update `this.state` synchronously, and in fact cause a re-render midway through whatever handler is running. Most of the time it's the former
behaviour, unless the call to `setState` comes from a context that React doesn't know about and isn't controlling (ie., not a `onX` handler on a component, and not a lifecycle method. stuff like timers).

This commit includes a minor change to Spindle's model initiation that works around this problem for the `diceroll` example by initiating `this.state.model` to `Model()` in the constructor, instead of setting it to `null`.

That new behaviour probably breaks spindle's nested/root detection via context, since it means that the first render will occur instead of being deferred until `componentDidMount`. Whoops :( need to add back another thing to see
if we've initialied yet.

I think `spindle` will need to start being just slightly more hands-on with managing the render context:

- initialization phase
  - manage context and root-vs-nested detection
  - initialize state and run `handleProps` to get any commands to start with
  - finally do the render with the initialized stuff

- update phase
  - manage calls to update(), and synchronous hand-off to commands
  - finally render with whatever new stuff we have

In both phases, I think it's best to manage `state` ourselves so we can always pass the most-recent version to whatever handler needs it. This will prevent mid-stream re-renders when we know the update context but react doesn't
(like async stuff coming from subscripitons or commands), and ensure we're always getting the state we expect. We can flush it once at the end with `setState` and be confident that we'll only do one render-per-update.

A possible alternative without managing state batching might be to hook `commands` into `componentDidUpdate`. This will mean that `render` will always run before the command runs (so if the command calls an action synchronously,
we'll do two synchronous renders), but any `setState` batching will be resolved before we run `command`, so it will see what it expects.