Explicitly manage state updates
React batches state updates, so reading `this.state.model` anywhere is unreliable -- it might be stale.

This commit keeps our own ref to the current state on the model instance that we can use in `update`, `subscriptions`, etc., that we know is up to date. It eagerly calls `setState`, which react will batch and not flush to the DOM in most cases, until the update stuff is all settled.

If we end up flushing to the DOM too often for stuff coming in from timers and things, then we can either:

- keep track of the update cycle ourselves in parallel to react and only call `setState` at the end, or
- see if we can hook into react somehow to give it context to start an update cycle